requirements.txt
biopython
prody
requests
joblib